refactor(api-gateway): Split dashboard HTTP/auth helpers out of dashboard.controller
Extracts authenticateGateway/trustedIdentityHeaders/fetchJson/postJson/
degrade into dashboard-http.helpers.ts, re-exported from
dashboard.controller.ts for pada-visits.controller.ts/padas.controller.ts's
existing imports. Addresses PR #163 review feedback (console.error already
uses pino logger; file was still 293 lines, over the repo's ~250-line guideline).
